% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnoC.R
\name{UnoC}
\alias{UnoC}
\title{Uno type C-index}
\usage{
UnoC(time, event, predicted, tau = NULL)
}
\arguments{
\item{time}{The vector of the observed times.}

\item{event}{The vector of events, 1 for event, 0 for censored.}

\item{predicted}{The predicted value varies depending on the model used. For a Cox model, the predicted value is the linear combination of the predictors. For a random survival forests model, the predicted value is the predicted mortality, which represents the number of events in the dataset if all observations had this set of predictors. For survival models using the \code{survivalmodels} package, the predicted value is the risk, defined as the rank of the negative mean survival time. Generally, a higher predicted value indicates worse survival.}

\item{tau}{The truncation point, the default is the largest event time.}
}
\description{
Uno type C-index
}
\examples{
# Example usage:
\donttest{
  library(survival)
  set.seed(123)
  time <- runif(10, 0, 100)
  event <- rbinom(10, 1, 0.5)
  predicted <- runif(10, 0, 1)
  UnoC(time, event, predicted, tau = 50)
}
}
